Move extensions classes to extension folder
(cherry picked from commit e18b3e9f73383c5cef15f979d0c930e377184dcb)
