+++ Kentico.Kontent.Management.Tests/ManagementClientTests/AssetTests.cs
@@ -10,6 +10,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
+using Kentico.Kontent.Management.Extenstions;
 using static Kentico.Kontent.Management.Tests.ManagementClientTests.Scenario;
 
 namespace Kentico.Kontent.Management.Tests.ManagementClientTests
