+++ Kentico.Kontent.Management.Tests/ManagementClientTests/ContentItemTests.cs
@@ -1,4 +1,5 @@
 ﻿using Kentico.Kontent.Management.Exceptions;
+using Kentico.Kontent.Management.Extenstions;
 using Kentico.Kontent.Management.Models.Items;
 using Kentico.Kontent.Management.Models.Shared;
 using System;
